Extract ImporterProperties from ImporterResidents
- Add Property.search_by_address, .not_test
